requirements.txt
Fabric
Flask-SQLAlchemy
SQLAlchemy-serializer
Flask-WTF
# WTForms
# coverage
# ecdsa
# paramiko
# pycrypto
